Separate render buffer into base class
