+++ Source/HelixToolkit.SharpDX.Shared/HelixToolkit.SharpDX.Shared.projitems
@@ -112,12 +112,14 @@
     <Compile Include="$(MSBuildThisFileDirectory)Model\ObservableObject.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Model\Collection\Vector2Collection.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Model\Collection\Vector3Collection.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Render\RenderBuffers\DX11RenderBufferBase.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Render\RenderBuffers\DX11SwapChainRenderBufferProxy.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Render\Renderer\ImmediateContextRenderer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Render\Renderer\DeferredContextRenderer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Render\RenderHost\RenderHostBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Render\RenderHost\DefaultRenderHost.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Render\DeviceContextProxy.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Render\DX11RenderBufferProxy.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Render\RenderBuffers\DX11Texture2DRenderBufferProxy.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Interface\IRenderer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Render\RenderContext.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Render\RenderContext2D.cs" />
