+++ Source/HelixToolkit.SharpDX.Shared/ShaderManager/EffectsManager.cs
@@ -2,7 +2,7 @@
 The MIT License (MIT)
 Copyright (c) 2018 Helix Toolkit contributors
 */
-#define DEBUGMEMORY
+//#define DEBUGMEMORY
 using System;
 using System.Collections.Generic;
